[refactor] 拆分 EmojiManager
